requirements.txt
hunspell==0.5.5
pronouncing==0.2.0
wordfreq==2.2.1
python-datamuse==1.2.1
